// src/app/modelos/tramite.interface.ts
export interface TramiteI{
    //================================================
    // ==== AQUI PONDRE MAS DATOS SI SE REQUIERE ====
    //================================================

    id_tramite: string;
    estado: string;
    nombre: string;
    apellido: string;
    codtramite: string;
    tipo: string;
    fechatramite: string;
    fecha_hito_1: string;
    fecha_hito_2: string;
    nro_orden: string;
    codigo: string;
    escuela: string;
    facultad: string;
    dni: string;
    sexo: string;
    verificacion: string;
    nrocelular: string;

    //
    mes: string;
    cantidad_tramites: string;
    //
    tipo_1: string;
    tipo_2: string;

 }
 